requirements.txt
slack-sdk>=3.26.1
requests>=2.31.0
Flask>=3.0.0
google-generativeai>=0.3.2
python-dotenv>=1.0.0
gunicorn>=21.2.0 